Move authorize into resource system
