+++ scrypto/src/component/mod.rs
@@ -6,4 +6,4 @@ mod system;
 pub use component::{Component, ComponentId, ComponentState, ParseComponentIdError};
 pub use lazy_map::{LazyMap, ParseLazyMapError};
 pub use package::{Package, PackageId, ParsePackageIdError};
-pub use system::{component_system, init_component_system, authorize, ComponentSystem};
+pub use system::{component_system, init_component_system, ComponentSystem};
