+++ scrypto/src/resource/mod.rs
@@ -28,5 +28,5 @@ pub use resource_flags::*;
 pub use resource_permissions::*;
 pub use resource_type::ResourceType;
 pub use supply::Supply;
-pub use system::{init_resource_system, resource_system, ResourceSystem};
+pub use system::{init_resource_system, resource_system, ResourceSystem, authorize};
 pub use vault::{ParseVaultError, Vault};
