+++ scrypto/src/resource/system.rs
@@ -72,6 +72,26 @@ pub fn resource_system() -> &'static mut ResourceSystem {
     unsafe { RESOURCE_SYSTEM.as_mut().unwrap() }
 }
 
+pub fn authorize<F,O>(bucket: &Bucket, func: F) -> O where F: FnOnce() -> O {
+    let input = CreateBucketProofInput {
+        bucket_id: bucket.0
+    };
+    let output: CreateBucketProofOutput = call_engine(CREATE_BUCKET_PROOF, input);
+
+    let input = PushOntoAuthWorktopInput { proof_id: output.proof_id };
+    let _: PushOntoAuthWorkTopOutput = call_engine(PUSH_ONTO_AUTH_WORKTOP, input);
+
+    let return_value = func();
+
+    let input = PopFromAuthWorktopInput { };
+    let output: PopFromAuthWorkTopOutput = call_engine(POP_FROM_AUTH_WORKTOP, input);
+
+    let input = DropProofInput { proof_id: output.proof_id };
+    let _: DropProofOutput = call_engine(DROP_PROOF, input);
+
+    return_value
+}
+
 /// This macro creates a `&ResourceDef` from a `ResourceDefId` via the
 /// Radix Engine resource subsystem.
 #[macro_export]
